Fixing an errant object reference in the settings
The last line of the file had a "s....." object that should have been "os...." akin to the line before it. Checking this one in before I forget because even though it's not the entire reason the docker image wouldn't compile it certainly is problematic for anything the app does that requires being auth'd into the associated Google account.